% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nhd_plus}
\alias{get_nhd_plus}
\title{function to download nhd files}
\usage{
get_nhd_plus(download_url, data_dir = normalizePath("nsink_data/"),
  download_again = FALSE)
}
\arguments{
\item{download_url}{url to download}

\item{data_dir}{The data dir}

\item{force}{Force new download}
}
\description{
function to download nhd files
}
\keyword{internal}
