% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_off_network_removal}
\alias{nsink_calc_off_network_removal}
\title{Calculates off network waterbody and stream nitrogen removal}
\usage{
nsink_calc_off_network_removal(
  input_data,
  off_network_lakes,
  off_network_streams,
  off_network_canalsditches
)
}
\arguments{
\item{input_data}{A named list with "streams", "lakes",
"network_removal", "tot", and "raster_template".}

\item{off_network_lakes}{Optional argument to set removal for waterbodies
that are not part of the hydrologic network in NHDPlus. Default value is to
use the 75th percentile of removal from other lakes in the HUC.  If another
value is desired provide a single numeric ranging from 0 to 1.}

\item{off_network_streams}{Optional argument to set removal for streams that
are not part of the hydrologic network in NHDPlus. Default value is to use
the median removal from first order streams in the HUC.  If another value
is desired provide a single numeric ranging from 0 to 1.}

\item{off_network_canalsditches}{Optional argument to set removal for canals
and ditches that are not part of the hydrologic network in NHDPlus. Default
value is to use the 25th percentile of removal from third order streams in
the HUC. If another value is desired provide a single numeric ranging from
0 to 1.}
}
\value{
raster and vectors of off network nitrogen removal
}
\description{
Calculates off network waterbody and stream nitrogen removal
}
\keyword{internal}
