% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_removal_type}
\alias{nsink_calc_removal_type}
\title{Create removal type raster}
\usage{
nsink_calc_removal_type(removal_rasters)
}
\arguments{
\item{removal_rasters}{A named list of "land_removal", "stream_removal,
"lake_removal", and "raster_template" rasters plus a
sf object "huc".}
}
\value{
raster of landscape nitrogen removal
}
\description{
Create removal type raster
}
\keyword{internal}
