% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_calc_total_removal}
\alias{nsink_calc_total_removal}
\title{Calculate total nitrogen removal}
\usage{
nsink_calc_total_removal(flowpath_vector)
}
\arguments{
\item{flowpath_vector}{a vector of removal values extracted from a nitrogen
removal raster generated by
\code{\link{nsink_extract_flowpath}}.}
}
\value{
A single numeric vector for total nitrogen removal along the flowpath
}
\description{
Nitrogen removal along a flowpath is determined by the types of landcover
encountered.  This function takes a vector of removal estimates and
calculates the total nitrogen removal for that flowpath.
}
\keyword{internal}
