% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_create_segment_ids}
\alias{nsink_create_segment_ids}
\title{Create ID's for unique values in a vector}
\usage{
nsink_create_segment_ids(x)
}
\arguments{
\item{x}{a vector of values to create unique ID's}
}
\value{
a vector of unique ID's
}
\description{
This functions takes a vector of values and creates a unique ID for adjacent
and identical values.
}
\keyword{internal}
