% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_create_summary}
\alias{nsink_create_summary}
\title{Create a nitrogen removal summary}
\usage{
nsink_create_summary(removal_vector, type_vector)
}
\arguments{
\item{removal_vector}{A flowpath vector of nitrogen removal generated by
\code{\link{nsink_extract_flowpath}}.}

\item{type_vector}{A flowpath vector of removal type generated by
\code{\link{nsink_extract_flowpath}}.}
}
\value{
a data frame summarizing nitrogen removal along a flowpath
}
\description{
This functions takes a nitrogen removal and removal type flowpath and creates
a data frame that summarizes the removal along that flowpath
}
\keyword{internal}
