% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_extract_flowpath}
\alias{nsink_extract_flowpath}
\title{Extract a vector raster cells of removal and the type of removal for a flowpath}
\usage{
nsink_extract_flowpath(flowpath, rasters)
}
\arguments{
\item{flowpath}{a flowpath to summarize nitrogen removal}

\item{rasters}{the removal or type rasters, generated by
\code{\link{nsink_calc_removal}}}
}
\value{
A vector of all cells along the flowpath with either removal amount
        or removal type (depending on input for the rasters argument).
}
\description{
This function takes a flowpath as input and extracts out the nitrogen removal
and they type of removal.
}
\keyword{internal}
