% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_flowpath.R
\name{nsink_generate_flowpath}
\alias{nsink_generate_flowpath}
\title{Generate and clean a flowpath for N-Sink}
\usage{
nsink_generate_flowpath(starting_location, input_data)
}
\arguments{
\item{starting_location}{An \code{\link{sf}} point location as a starting
point for the flowpath.  Projection must match
projection in input_data}

\item{input_data}{A list of input data with (at least) "fdr", "streams" and
"tot", and "raster_template". These may be generated with
\code{\link{nsink_prep_data}}.}
}
\value{
An \code{\link{sf}} LINESTRING object of the flowpath that starts at
        the \code{starting_location} and ends at the ouflow of the HUC.
}
\description{
This function takes an XY location as a starting point and generates a hybrid
(flow direction derived flowpath on land plus NHDPlus vector downstream
flowpath) flowpath for use in the N-Sink nitrogen removal analysis.
}
\examples{
\dontrun{
library(nsink)
library(sf)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea)
removal_rasters <- nsink_calc_removal(niantic_nsink_data)
pt <- c(1948121,2295822)
start_loc <- st_sf(st_sfc(st_point(c(pt)), crs = aea))
fp <- nsink_generate_flowpath(start_loc, niantic_nsink_data)
}
}
