% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsink_get_closest}
\alias{nsink_get_closest}
\title{Get closest}
\usage{
nsink_get_closest(v1, v2)
}
\arguments{
\item{v1}{The first vector of values, likely length or area of a feature}

\item{v2}{The second vector of values, also likely length or area of a
feature.}
}
\value{
Returns a vector, of length v1, of the index from v2 that is closest
        in absolute value to each value in v1.
}
\description{
This function will return the index of one vector that is closest, by
absolute values, to the values in another vector.  Usually used to
identify the closest lake or stream, in area or length respectively, to
another lake or stream.
}
\keyword{internal}
