% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_get_data.R
\name{nsink_get_data}
\alias{nsink_get_data}
\title{Get's N-Sink data for a given HUC}
\usage{
nsink_get_data(huc, data_dir = normalizePath("nsink_data"),
  force = FALSE)
}
\arguments{
\item{huc}{A character string of a HUC12 identifier.  Currently can run only a single HUC at a time.}

\item{data_dir}{A directory to store N-Sink data downloads.  Defaults to
"nsink_data" inside of the current working directory.
Created if it doesn't exist.}

\item{force}{Logical to determine if files should be downloaded
again if they already exist locally.}
}
\value{
Returns a list with the huc used and the directory where the data is
        stored.
}
\description{
The required datasets for the N-sink analysis are available from multiple,
online resources.  This function takes a HUC as input and downloads local
copies of those datasets.
}
\examples{
\dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
nsink_get_data(huc = niantic_huc)
}
}
