% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_flowpath.R
\name{nsink_get_flowline}
\alias{nsink_get_flowline}
\title{Get flowlines that intersect with a flowpath}
\usage{
nsink_get_flowline(flowpath_ends, streams)
}
\arguments{
\item{flowpath_ends}{An \code{sf} LINESTRING of the flowpath ends, generated
with \code{\link{nsink_get_flowpath_ends}}}

\item{streams}{NHDPlus streams from}
}
\value{
an \code{sf} object of the NHDPlus flowlines that occur after a
        raster flowpath intersects the stream network.
}
\description{
Extract flowlines that intersect with flowpath ends.  This uses the actual
flowlines as a part for the flowpath instead of simply using the raster
derived flowpaths wich do not follow the flowlines exactly.
}
\keyword{internal}
