% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_flowpath.R
\name{nsink_get_flowpath_ends}
\alias{nsink_get_flowpath_ends}
\title{Get flowpath beginning and ends}
\usage{
nsink_get_flowpath_ends(flowpath, streams)
}
\arguments{
\item{flowpath}{An \code{sf} LINESTRING of the flowpath, generated with \code{\link{nsink_generate_flowpath}}}

\item{streams}{NHDPlus streams from}
}
\value{
An \code{sf} object of the portions of the flowpath that are not
        represented by the NHDPlus flowlines
}
\description{
Flowpath from land is only portion that needs to be generated from the flow
direction grid.  This function extracts those portions of the generated
flowpath.  There may be issues with off network waterbodies...  May need to
find all sections without connected flowpaths...
}
\keyword{internal}
