% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_get_data.R
\name{nsink_get_huc_id}
\alias{nsink_get_huc_id}
\title{Look up HUC 12 ID from a HUC name}
\usage{
nsink_get_huc_id(huc_name)
}
\arguments{
\item{huc_name}{character string of a HUC Name}
}
\value{
A data frame with HUC_12 and HU_12_NAME that match the huc_name
}
\description{
This function takes a HUC Name and returns all possible matching HUC 12 IDs.
}
\examples{
nsink_get_huc_id(huc_name = "Niantic River")
}
