% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_get_plus_remotepath.R
\name{nsink_get_plus_remotepath}
\alias{nsink_get_plus_remotepath}
\title{Get remote path for NHDPlus components}
\usage{
nsink_get_plus_remotepath(vpu, component = c("NHDSnapshot", "FdrFac",
  "EROMExtension", "WBDSnapshot", "NHDPlusAttributes"))
}
\description{
Code modified from https://github.com/jsta/nhdR/blob/master/R/utils.R.  Still
need to figure out best way to acknowledge jsta as author and include GPL
}
