% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_group_land_off_network}
\alias{nsink_group_land_off_network}
\title{nsink group land off network}
\usage{
nsink_group_land_off_network(land_removal_df)
}
\arguments{
\item{land_removal_df}{land and off network summary df}
}
\description{
nsink group land off network
}
\keyword{internal}
