% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_load.R
\name{nsink_load}
\alias{nsink_load}
\title{Load an existing N-Sink analysis folder}
\usage{
nsink_load(input_folder, base_name = "nsink_")
}
\arguments{
\item{input_folder}{Folder to that contains nsink files produced by
\code{\link(nsink_build)}}

\item{base_name}{a base name used to assign objects to the global environment.}
}
\value{
Creates several lists in the global environment that would normally
        be created when running an N-Sink analysis.  These include:
        a \code{\link{nsink_prep_data}} object,
        a \code{\link{nsink_calc_removal}} object, and a
        \code{\link{nsink_generate_static_maps}} object
}
\description{
Load an existing N-Sink analysis folder
}
\examples{
\dontrun{
library(nsink)
#' library(nsink)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
nsink_build(nsink_get_huc_id("Niantic River")$huc_12, aea,
            output_folder = "nsink_output", samp_dens = 300)
nsink_load(input_folder = "nsink_output",
           basename = "nsink_")
}
}
