% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_merge_removal}
\alias{nsink_merge_removal}
\title{Merges removal rasters into single raster}
\usage{
nsink_merge_removal(removal_rasters)
}
\arguments{
\item{removal_rasters}{A named list of "land_removal", "stream_removal",
"lake_removal", and "raster_template" rasters plus a
sf object "huc".}
}
\value{
Raster of landscape nitrogen removal
}
\description{
Merges removal rasters into single raster
}
\keyword{internal}
