% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_plot.R
\name{nsink_plot}
\alias{nsink_plot}
\title{N-Sink plot function for the a list of nsink static maps}
\usage{
nsink_plot(static_maps, map = c("removal", "transport", "delivery"))
}
\arguments{
\item{static_maps}{A list of \code{raster} objects that can be created
via \code{\link{nsink_generate_static_maps}}.  The list
should have removal_effic, delivery_idx, and
transport_idx.}

\item{map}{A character of either, "removal", "transport", or "delivery."}
}
\description{
This function creates a simple plot with pre-selected color palettes from a
list of static maps created by \code{\link{nsink_generate_static_maps}}.
This is meant as a quick means to visualize the various static maps.
}
\examples{
\dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc, data_dir = "nsink_data")
aea <- 5072
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea,
                                      data_dir = "nsink_data")
removal <- nsink_calc_removal(niantic_nsink_data)
static_maps <- nsink_generate_static_maps(niantic_nsink_data, removal,
samp_dens = 900)
nsink_plot(static_maps, "transport")
}
}
