% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_plot.R
\name{nsink_plot_delivery}
\alias{nsink_plot_delivery}
\title{N-Sink plot function for the nitrogen delivery index}
\usage{
nsink_plot_delivery(
  delivery_idx,
  breaks = c(20, 40, 60, 80, 100),
  colors = c("#FFBEBE", "#F57A7A", "#A80405", "#652600")
)
}
\arguments{
\item{delivery_idx}{A delivery index \code{raster} most likely created
via \code{\link{nsink_generate_static_maps}}.}

\item{breaks}{A vector of values specifying breakpoints to break the
delivery \code{raster}.  Must be one more than number of
colors.}

\item{colors}{A vector of hexcodes for the colors}
}
\description{
This function creates a simple plot with a pre-selected color palette,
although different breaks and colors can be provided by the user.  This is
meant as a quick means to visualize delivery index.
}
