% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_plot.R
\name{nsink_plot_removal}
\alias{nsink_plot_removal}
\title{N-Sink plot function for removal efficiency}
\usage{
nsink_plot_removal(
  removal_effic,
  breaks = c(0.2, 0.4, 0.6, 0.8),
  colors = c("#D3FFBE", "#70A800", "#267300")
)
}
\arguments{
\item{removal_effic}{A removal efficiency \code{raster} that can be created
via \code{\link{nsink_generate_static_maps}}.}

\item{breaks}{A vector of values specifying breakpoints to break the
removal \code{raster}.  Must be one more than number of
colors.}

\item{colors}{A vector of hexcodes for the colors}
}
\description{
This function creates a simple plot with a pre-selected color palette,
although different breaks and colors can be provided by the user.  This is
meant as a quick means to visualize removal efficiency.
}
\keyword{internal}
