% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_plot.R
\name{nsink_plot_transport}
\alias{nsink_plot_transport}
\title{N-Sink plot function for the nitrogen transport index}
\usage{
nsink_plot_transport(
  transport_idx,
  breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100),
  colors = c("#38A1D0", "#7AB4C0", "#A2C8B0", "#C3DC9D", "#E2F088", "#F5E871",
    "#F9C159", "#F99844", "#F56A2E", "#EF2820")
)
}
\arguments{
\item{transport_idx}{A transport index \code{raster} most likely created
via \code{\link{nsink_generate_static_maps()}}.}

\item{breaks}{A vector of values specifying breakpoints to break the
transport \code{raster}.  Must be one more than number of
colors.}

\item{colors}{A vector of hexcodes for the colors}
}
\description{
This function creates a simple plot with a pre-selected color palette,
although different breaks and colors can be provided by the user.  This is
meant as a quick means to visualize removal efficiency.
}
