% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_fdr}
\alias{nsink_prep_fdr}
\title{Prepare flow direction data for N-Sink}
\usage{
nsink_prep_fdr(huc_sf, huc_raster, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{huc_raster}{A raster object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may be
downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
returns a raster object of the flow direction for the huc_sf but in
        the original fdr projection
}
\description{
Standardizes flow direction data by transforming data, clipping to HUC, ...
}
\keyword{internal}
