% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_lakes}
\alias{nsink_prep_lakes}
\title{Prepare lakes data for N-Sink}
\usage{
nsink_prep_lakes(huc_sf, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may be
downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
An sf object of the NHDPlus lakes for the huc_sf
}
\description{
Standardizes lakes data by transforming data, clipping to HUC, and
renaming columns.
}
\keyword{internal}
