% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_nlcd}
\alias{nsink_prep_nlcd}
\title{Prepare NLCD data for N-Sink}
\usage{
nsink_prep_nlcd(huc_sf, huc_raster, data_dir, year)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{huc_raster}{A raster object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}

\item{year}{The year of the nlcd and impervious data that was retrieved with
FedData's, \code{\link{get_nlcd}} function.}
}
\value{
A raster object of the NLCD for the huc_sf
}
\description{
Standardizes NLCD data by by projecting to the coorect CRS,
and standardizing file name.
}
\keyword{internal}
