% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_q}
\alias{nsink_prep_q}
\title{Prepare flow data for N-Sink}
\usage{
nsink_prep_q(data_dir)
}
\arguments{
\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
A tibble of the flow data
}
\description{
Standardizes flow data from the EROM tables.
}
\keyword{internal}
