% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_ssurgo}
\alias{nsink_prep_ssurgo}
\title{Prepare SSURGO data for N-Sink}
\usage{
nsink_prep_ssurgo(huc_sf, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
returns a sf object of the SSURGO data with hydric data added.
        for the huc_sf
}
\description{
Standardizes impervious data by transforming data, clipping to HUC, ...
}
\keyword{internal}
