% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_streams}
\alias{nsink_prep_streams}
\title{Prepare streams data for N-Sink}
\usage{
nsink_prep_streams(huc_sf, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
returns an sf object of the NHDPlus streams for the huc_sf
}
\description{
Standardizes streams data by transforming data, clipping to HUC, ...
}
\keyword{internal}
