% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_tot}
\alias{nsink_prep_tot}
\title{Prepare time of travel data for N-Sink}
\usage{
nsink_prep_tot(data_dir)
}
\arguments{
\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
returns a tibble of the time of travel data
}
\description{
Standardizes time of travel from the VAA tables.
}
\keyword{internal}
