% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_remove_openwater}
\alias{nsink_remove_openwater}
\title{Remove open water portions of the HUC}
\usage{
nsink_remove_openwater(huc_sf, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
An sf object of the HUC without salt water/open water area.
}
\description{
Uses SSURGO polys and the SSURGO musym = Ws to check for and remove any
portion of a HUC that is actually salt water.  This should account for the
coastal HUCs with large areas of open water.
}
\keyword{internal}
