% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsink_run_7z}
\alias{nsink_run_7z}
\title{Finds 7-zip}
\usage{
nsink_run_7z(zipfile, destdir, extract_again = FALSE)
}
\arguments{
\item{zipfile}{The zipfile to be extracted}

\item{destdir}{Where to put the extracted files}

\item{force}{Whether or not to extract again if the destination files
already exist}
}
\description{
This code is modified from https://github.com/jsta/nhdR/blob/master/R/utils.R
and https://github.com/jsta/nhdR/blob/master/R/get.R to determine if 7 zip is
available.  If available it unzips a 7z zipfile to a destination directory.
This avoids needing to use archive package which is only available via
GitHub.
}
\author{
Jemma Stachelek, \email{jsta@lanl.gov}
}
\keyword{internal}
