% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_flowpath.R
\name{nsink_split_flowline}
\alias{nsink_split_flowline}
\title{Split flowlines where they intersect with a flowpath}
\usage{
nsink_split_flowline(flowpath_ends, flowpath_network)
}
\arguments{
\item{flowpath_ends}{The ends of the flowpath that are not a part of the
network}

\item{flowpath_network}{The flowpath network}
}
\value{
An \code{sf} object of the NHDPlus flowlines split where the fp ends
        intersect with the flowlines.
}
\description{
Takes a flowpath input
}
\keyword{internal}
