% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_build.R
\name{nsink_write_static_maps}
\alias{nsink_write_static_maps}
\title{Write static maps to files}
\usage{
nsink_write_static_maps(static_maps, output_dir)
}
\arguments{
\item{static_maps}{A list of static maps, as output by
\code{\link{nsink_generate_static_maps}}}

\item{output_dir}{Output folder in which to save static maps as .tif}
}
\description{
Writes out static mapps as tiffs
}
\keyword{internal}
