% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_build.R
\name{nsink_build}
\alias{nsink_build}
\title{Build out required datasets for N-Sink}
\usage{
nsink_build(
  huc,
  projection,
  output_folder = normalizePath("nsink_output", winslash = "/"),
  data_dir = NULL,
  force = FALSE,
  samp_dens = 300,
  ...
)
}
\arguments{
\item{huc}{A character with the 12 digit HUC ID.  Maybe searched with
\code{\link{nsink_get_huc_id}}}

\item{projection}{Projection to use for all spatial data, passed as either an
EPSG code (as numeric) or WKT as string.}

\item{output_folder}{Folder to store downloaded data (if not in a separate
data folder) and write out processed nsink files.
Currently, the processed files will be overwritten if
the same output folder is used.  To run different
HUC12's specify separate output folders.}

\item{data_dir}{An optional folder to hold downloaded data.  The same data
directory can be used to hold data for multiple HUCs.  Data
will not be downloaded again if it already exists in this
folder.}

\item{force}{Logical value used to force a new download if data already
exists on file system}

\item{samp_dens}{The \code{samp_dens} controls the density of points to use when
creating the nitrogen removal heat map.  The area of the
watershed is sampled with points that are separated by the
\code{samp_dens} value.  The larger the value, the fewer the points.}

\item{...}{Passes to \code{\link{nsink_calc_removal}} for the off network
arguments: \code{off_network_lakes}, \code{off_network_streams},
and \code{off_network_canalsditches}.}
}
\value{
a list providing details on the huc used and the output location of
        the dataset
}
\description{
This function is a wrapper around the other functions and runs all of those
required to build out the full dataset need for a huc plus develop the four
static N-Sink maps: the nitrogen loading index, nitrogen removal effeciency,
nitrogen transport index, and the nitrogen delivery index.  The primary
purpose of this is to use the nsink package to develop the required datasets
for an nsink application to be built in other software.  This will take some
time to complete as it is downloading 500-600 Mb of data, processing that data
and then creating output files.
}
\examples{
\dontrun{
library(nsink)
aea <- 5072
nsink_build(nsink_get_huc_id("Niantic River")$huc_12, aea,
            output_folder = "nsink_output", data_dir = "nsink_data",
             samp_dens = 600)
}
}
