% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_build.R
\name{nsink_build}
\alias{nsink_build}
\title{Build out required datasets for N-Sink}
\usage{
nsink_build(huc, projection,
  output_folder = normalizePath("nsink_output"), force = FALSE,
  fact = 300)
}
\arguments{
\item{huc}{A character with the 12 digit HUC ID.  Maybe searched with
\code{\link{nsink_get_huc_id}}}
}
\value{
a list providing details on the huc used and the output location of
        the dataset
}
\description{
This function is a wrapper around the other functions and runs all of those
required to build out the full dataset need for a huc plus develop the four
static N-Sink maps: the nitrogen loading index, nitrogen removal effeciency,
nitrogen transport effeciencey, and the nitrogen delivery index.  The primary
purpose of this is to use the nsink package to develop the require datasets
for an nsink application to be built in other software.  This will take some
time to complete as it is downloaing 500-600 Mb of data, processing that data
and then creating output files.
}
\examples{
library(nsink)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
nsink_build(nsink_get_huc_id("Niantic River")$huc_12, aea)
}
