% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_build.R
\name{nsink_build}
\alias{nsink_build}
\title{Build out required datasets for N-Sink}
\usage{
nsink_build(
  huc,
  projection,
  output_folder = normalizePath("nsink_output"),
  force = FALSE,
  fact = 300,
  signal_finish = FALSE
)
}
\arguments{
\item{huc}{A character with the 12 digit HUC ID.  Maybe searched with
\code{\link{nsink_get_huc_id}}}

\item{projection}{Projection to use for all spatial data, passed as a PROJ
string}

\item{output_folder}{Folder to store downloaded data and process nsink files}

\item{force}{Logical value used to force a new download if data already
exists on file system}

\item{fact}{The \code{fact} controls the density of points to use when
creating the nitrogen removal heat map.  The area of the
watershed is sampled with points that are separated by the
\code{fact} value.  The larger the value, the fewer the points.}

\item{signal_finish}{Logical that triggers a sound when build is finished.}
}
\value{
a list providing details on the huc used and the output location of
        the dataset
}
\description{
This function is a wrapper around the other functions and runs all of those
required to build out the full dataset need for a huc plus develop the four
static N-Sink maps: the nitrogen loading index, nitrogen removal effeciency,
nitrogen transport effeciencey, and the nitrogen delivery index.  The primary
purpose of this is to use the nsink package to develop the required datasets
for an nsink application to be built in other software.  This will take some
time to complete as it is downloaing 500-600 Mb of data, processing that data
and then creating output files.
}
\examples{
\dontrun{
library(nsink)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
nsink_build(nsink_get_huc_id("Niantic River")$huc_12, aea,
            output_folder = "nsink_output", fact = 300)}
}
