% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_lake_removal}
\alias{nsink_calc_lake_removal}
\title{Calculates lake-based nitrogen removal}
\usage{
nsink_calc_lake_removal(input_data)
}
\arguments{
\item{input_data}{A named list with "streams", "lakes", "tot", "lakemorpho",
and "raster_template".}
}
\value{
raster of lake based nitrogen removal
}
\description{
Calculates lake-based nitrogen removal
}
\keyword{internal}
