% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_land_removal}
\alias{nsink_calc_land_removal}
\title{Calulates land-based nitrogen removal}
\usage{
nsink_calc_land_removal(input_data)
}
\arguments{
\item{input_data}{A named list with "ssurgo", "impervious", and
"raster_template".}
}
\value{
raster of land based nitrogen removal
}
\description{
Calulates land-based nitrogen removal
}
\keyword{internal}
