% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_land_removal}
\alias{nsink_calc_land_removal}
\title{Calculates land-based nitrogen removal}
\usage{
nsink_calc_land_removal(input_data)
}
\arguments{
\item{input_data}{A named list with "ssurgo", "impervious", "lakes", and
"raster_template".}
}
\value{
list with raster and vector versions of land based nitrogen removal
}
\description{
Calculates land-based nitrogen removal
}
\keyword{internal}
