% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_removal}
\alias{nsink_calc_removal}
\title{Calculates N-Sink nitrogen removal percent}
\usage{
nsink_calc_removal(input_data, method = c("raster", "hybrid"))
}
\arguments{
\item{input_data}{A list of input datasets created with
\code{\link{nsink_prep_data}}.}

\item{method}{The type of data used for summarizing removal along the flowpath.
Raster data is extracted with a flowpath for the entire length of
the flowpath.  If hybrid is selected, only the land portion uses
raster data, the remainder of the flowpath is summarized based on
the vector streamlines. Default is raster}
}
\value{
If the rater method is used,  a raster stack with one layer with
        nitrogen removal, a second layer with the type of removal (e.g.
        hydric soils, lakes, streams), and a third layer with a clean raster
        template is returned.  If the hybrid method is used a list is
        returned with land_removal, stream_removal, and lake removal.
}
\description{
Starting with base data layers of NHDPlus, SSURGO, impervious surface, flow
velocity, and time of travel, this function calculates percentage of Nitrogen
removal.  Details for nitrogen removal calculation are from
\href{https://doi.org/10.1016/j.ecoleng.2010.02.006}{Kellogg et al. (2010)}.
This function assumes data has been downloaded with
\code{\link{nsink_get_data}} and has been prepared with
\code{\link{nsink_prep_data}}.
}
\examples{
\dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea)
nsink_calc_removal(niantic_nsink_data)
}
}
\references{
Kellogg, D. Q., Gold, A. J., Cox, S., Addy, K., & August, P. V.
            (2010). A geospatial approach for assessing denitrification sinks
            within lower-order catchments. Ecological Engineering, 36(11),
            1596-1606.
            \href{https://doi.org/10.1016/j.ecoleng.2010.02.006}{Link}
}
