% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_stream_removal}
\alias{nsink_calc_stream_removal}
\title{Calculates stream-based nitrogen removal}
\usage{
nsink_calc_stream_removal(removal_rasters)

nsink_calc_stream_removal(removal_rasters)
}
\arguments{
\item{removal_rasters}{list of land, stream, and lake nitrogen removal
rasters.}

\item{input_data}{list of data ...}
}
\value{
raster of stream based nitrogen removal

raster of stream based nitrogen removal
}
\description{
Calculates stream-based nitrogen removal

Merges removal rasters into single raster
}
\keyword{internal}
