% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_calc_removal.R
\name{nsink_calc_stream_removal}
\alias{nsink_calc_stream_removal}
\title{Calculates stream-based nitrogen removal}
\usage{
nsink_calc_stream_removal(input_data)
}
\arguments{
\item{input_data}{A named list with "streams", "q", "tot", and
"raster_template".}
}
\value{
raster and vector versions of stream based nitrogen removal
}
\description{
Calculates stream-based nitrogen removal
}
\keyword{internal}
