% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_create_summary}
\alias{nsink_create_summary}
\title{Create a nitrogen removal summary}
\usage{
nsink_create_summary(land_removal, network_removal)
}
\arguments{
\item{land_removal}{A data frame of land based nitrogen removal via the
generated flowpath and hydric removal raster.}

\item{network_removal}{A data frame of stream network nitrogen removal via
calculated stream and lake removal}
}
\value{
a data frame summarizing nitrogen removal along a flowpath
}
\description{
This functions takes a nitrogen removal and removal type data frame for land
and the flowpath network  and creates a data frame that summarizes the
removal along that flowpath.
}
\keyword{internal}
