% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsink_fix_data_directory}
\alias{nsink_fix_data_directory}
\title{Fix the data directory}
\usage{
nsink_fix_data_directory(data_dir)
}
\arguments{
\item{data_dir}{the data directory}
}
\value{
a string with the normalized path
}
\description{
This funciton takes the data directory and checks for existence, creates it
if it doesn't exist, then adds a trailing slash and normalizes the path for
the operating system
}
\keyword{internal}
