% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_static_maps.R
\name{nsink_generate_n_loading_index}
\alias{nsink_generate_n_loading_index}
\title{Generates the Nitrogen Loading Index}
\usage{
nsink_generate_n_loading_index(input_data)
}
\arguments{
\item{input_data}{list of input data}
}
\description{
This function reclassifies the NLCD data to an index of Nitrogen loading.
The index ranges from 0 to 1.
}
\keyword{internal}
