% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_static_maps.R
\name{nsink_generate_n_removal_heatmap}
\alias{nsink_generate_n_removal_heatmap}
\title{Generate Nitrogen Removal Heatmap}
\usage{
nsink_generate_n_removal_heatmap(
  input_data,
  removal,
  samp_dens,
  ncpu = future::availableCores() - 1
)
}
\arguments{
\item{input_data}{A list of input datasets created with
\code{\link{nsink_prep_data}}.}

\item{removal}{the removal raster stack or removal list, generated by
\code{\link{nsink_calc_removal}}}

\item{samp_dens}{A value, in the units of the input data, divided by total area of
the input HUC.  It is used to determine the number of points,
determined through a regular sample, to calculate removal.  For
instance a value of 90 would roughly equate to a point per every
90 meters.}

\item{ncpu}{number of CPUs to use for calculating flowpath removal for larger
(i.e. greater than 50) number of flowpaths.  Defaults to one
minus the number of cores available.}
}
\description{
Generates the heatmap
}
\keyword{internal}
