% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_generate_static_maps.R
\name{nsink_generate_static_maps}
\alias{nsink_generate_static_maps}
\title{Generate N-Sink Static Maps for a given HUC}
\usage{
nsink_generate_static_maps(input_data, removal,
  data_dir = normalizePath("nsink_data/"), custom_load = NULL)
}
\arguments{
\item{input_data}{A list of input datasets created with
\code{\link{nsink_prep_data}}.}

\item{removal}{the removal raster stack or removal list, generated by
\code{\link{nsink_calc_removal}}}

\item{data_dir}{Base directory that contains N-Sink data folders and prepped
data.  Data may be downloaded with the
\code{\link{nsink_get_data}} function and prepared with the
\code{\link{nsink_prep_data}} function. A new folder named
"static_maps" with the .tif files will be created in this
folder.}

\item{custom_load}{Place holder for a custom loading index.  Currently uses
values culled from the literature.  Custom loads are not yet
implemented}
}
\description{
Generate N-Sink Static Maps for a given HUC
}
\examples{
\dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea)
niantic_removal <- nsink_calc_removal(niantic_nsink_data)
nsink_generate_static_maps(niantic_nsink_data, niantic_removal, "palmer_nsink_data/")
}
}
