% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_get_data.R
\name{nsink_get_data}
\alias{nsink_get_data}
\title{Get's N-sink data for a given HUC}
\usage{
nsink_get_data(location, data_dir = paste0(getwd(), "/nsink_data"))
}
\arguments{
\item{location}{An sf polygon object of the HUC (or really any spatial object) for
which you wish to build the N-Sink database.}

\item{data_dir}{A directory to store N-Sink data downloads.  Defaults to
current working directory.}
}
\description{
The required datasets for the N-sink analysis are available from multiple,
online resources.  This function takes a HUC as input and downloads local
copies of those datasets.
}
\examples{
data(niantic_huc)
nsink_get_data(location = niantic_huc)
}
