% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsink_get_plus_remotepath}
\alias{nsink_get_plus_remotepath}
\title{Get remote path for NHDPlus components}
\usage{
nsink_get_plus_remotepath(
  rpu,
  component = c("NHDSnapshot", "FdrFac", "EROMExtension", "WBDSnapshot",
    "NHDPlusAttributes")
)
}
\arguments{
\item{rpu}{raster processing unit for NHDPlus. available in nsink:::wbd_lookup}

\item{component}{which component to download}
}
\description{
Code modified from https://github.com/jsta/nhdR/blob/master/R/utils.R.  Still
need to figure out best way to acknowledge jsta as author and include GPL
}
\keyword{internal}
