% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_prep_data.R
\name{nsink_prep_nlcd}
\alias{nsink_prep_nlcd}
\title{Prepare NLCD data for N-Sink}
\usage{
nsink_prep_nlcd(huc_sf, huc_raster, data_dir)
}
\arguments{
\item{huc_sf}{An sf object of the Watershed Boundaries Dataset HUC12}

\item{huc_raster}{A raster object of the Watershed Boundaries Dataset HUC12}

\item{data_dir}{Base directory that contains N-Sink data folders.  Data may
be downloaded with the \code{\link{nsink_get_data}} function.}
}
\value{
returns a raster object of the NLCD for the huc_sf
}
\description{
Standardizes NLCD data by transforming data, clipping to HUC, ...
}
\keyword{internal}
