% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_summarize_flowpath}
\alias{nsink_summarize_flowpath}
\title{Summarize nitrogen removal along a flowpath}
\usage{
nsink_summarize_flowpath(
  flowpath,
  removal,
  method = c("raster", "hybrid"),
  filter_window = 3
)
}
\arguments{
\item{flowpath}{a flowpath to summarize nitrogen removal}

\item{removal}{the removal raster stack or removal list, generated by
\code{\link{nsink_calc_removal}}}

\item{method}{a character string which determines if the caclulations are
done using the raster method or a hybrid method
(not currently implemented).}

\item{filter_window}{A numeric value indicating the size of the filter
window. Default value is 3.  See argument description
for \code{width} from \code{\link{rollmax}}. Only used
if the raster method is chosen.}
}
\value{
A data frame is returned that with a summary of nitrogen removal
}
\description{
Nitrogen removal varies along a flowpath as they include different land cover
and waterbody types that have different nitrogen reduction capabilities.
This function takes a flowpath generated by
\code{\link{nsink_generate_flowpath}} as required input and returns and
estimate of total flow path removal as well as removal by type.
}
\examples{
\dontrun{
library(nsink)
library(sf)
library(lwgeom)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc)
aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea)
removal <- nsink_calc_removal(niantic_nsink_data)
pt <- c(1948121,2295822)
start_loc <- st_sf(st_sfc(st_point(c(pt)), crs = aea))
fp <- nsink_generate_flowpath(start_loc, niantic_nsink_data)
nsink_summarize_flowpath(fp, removal)
pt <- c(1956582, 2287697)
start_loc <- st_sf(st_sfc(st_point(c(pt)), crs = aea))
fp <- nsink_generate_flowpath(start_loc, niantic_nsink_data,
                                     method = "hybrid")
nsink_summarize_flowpath(fp, removal, method = "hybrid")
}
TODO Raster method: Deal with adjacent hydrics, prior to entering stream
     network with different removal by combining and averaging?  If I get
     hybrid method working might not need to do this.
TODO Raster method: Remove hydric that occurs once flowpath enters the stream
     network. Replace with values from either side. If I get hybrid method
     working might not need to do this.

}
