% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsink_summarize_flowpath.R
\name{nsink_summarize_flowpath}
\alias{nsink_summarize_flowpath}
\title{Summarize nitrogen removal along a flowpath}
\usage{
nsink_summarize_flowpath(flowpath, removal)
}
\arguments{
\item{flowpath}{a flowpath to summarize nitrogen removal}

\item{removal}{the removal raster stack or removal list, generated by
\code{\link{nsink_calc_removal}}}
}
\value{
A data frame is returned that with a summary of nitrogen removal.  The
}
\description{
Nitrogen removal varies along a flowpath as they include different land cover
and waterbody types that have different nitrogen reduction capabilities.
This function takes a flowpath generated by
\code{\link{nsink_generate_flowpath}} as required input and returns and
estimate of total flow path removal as well as removal by type.
}
\examples{
\dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc, data_dir = "nsink_data")
aea <- "+proj=aea +lat_0=23 +lon_0=-96 +lat_1=29.5 +lat_2=45.5 +x_0=0 +y_0=0
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea,
                                      data_dir = "nsink_data")
removal <- nsink_calc_removal(niantic_nsink_data)
pt <- c(1948121, 2295822)
start_loc <- st_sf(st_sfc(st_point(c(pt)), crs = aea))
fp <- nsink_generate_flowpath(start_loc, niantic_nsink_data)
flow_summary <- nsink_summarize_flowpath(fp, removal)
flow_summary
}
}
